+++ src/reducers/wizard.js
@@ -5,7 +5,9 @@ const initial = Map({
   type: null,
   mnemonicConfirmed: false,
   showAccountManagement: false,
-  requiresMnemonic: false
+  requiresMnemonic: false,
+  showCreateUsername: false,
+  showIdentityManagement: false
 });
 
 export default (state = initial, action) => {
@@ -21,7 +23,9 @@ export default (state = initial, action) => {
         isHidden: true,
         showAccountManagement: false,
         showSend: false,
-        showReceive: false
+        showReceive: false,
+        showIdentityManagement: false,
+        showCreateUsername: false
       });
 
     case "IMPORT_WALLET":
@@ -50,6 +54,18 @@ export default (state = initial, action) => {
 
     case "CLOSE_RECEIVE":
       return state.set("showReceive", false);
+
+    case "SHOW_CREATE_USERNAME":
+      return state.set("showCreateUsername", true);
+
+    case "HIDE_CREATE_USERNAME":
+      return state.set("showCreateUsername", false);
+
+    case "SHOW_IDENTITY_MANAGEMENT":
+      return state.set("showIdentityManagement", true);
+
+    case "HIDE_IDENTITY_MANAGEMENT":
+      return state.set("showIdentityManagement", false);
 
     default:
       return state;