+++ src/selectors/index.js
@@ -12,7 +12,7 @@ export const account = state =>
 export const names = state => account(state).get("identityIdByName", Map());
 export const accounts = state => wallet(state).get("accounts");
 export const selectedAccount = state => wallet(state).get("selectedAccount");
-
+export const identities = state => account(state).get("identities");
 export const username = state => account(state).get("selectedName");
 export const identityId = state => account(state).get("selectedIdentityId");
 export const balance = state => account(state).get("balance", Map());
@@ -31,14 +31,20 @@ const accountStatusSelectors = [
 
 export const accountStatus = createSelector(accountStatusSelectors, (...r) => {
   const [wizard, wallet, identityId, account, balance, names] = r;
-  console.log(names.size);
 
   if (wizard.get("showAccountManagement")) return "ACCOUNT_MANAGEMENT";
   if (wizard.get("showSend")) return "SEND";
+  if (wizard.get("showIdentityManagement")) return "IDENTITY_MANAGEMENT";
+  if (wizard.get("showCreateUsername")) return "CREATE_USERNAME";
   if (wizard.get("showReceive")) return "RECEIVE";
   if (wizard.get("isHidden")) return "HIDDEN";
 
-  // if (typeof account.selected !== "number") return "SELECT_ACCOUNT";
+  if (
+    wallet.get("accounts").size > 0 &&
+    typeof wallet.get("selectedAccount") !== "number"
+  ) {
+    return "ACCOUNT_MANAGEMENT";
+  }
 
   if (!wizard.get("mnemonicConfirmed") && !wizard.get("type")) {
     return "SELECT_WIZARD_TYPE";
@@ -59,6 +65,9 @@ export const accountStatus = createSelector(accountStatusSelectors, (...r) => {
   if (!identityId && total > 0 && names.size === 0) {
     return "CREATE_USERNAME";
   }
+  if (!identityId && total > 0 && names.size > 0) {
+    return "IDENTITY_MANAGEMENT";
+  }
 
   return "HIDDEN";
 });
@@ -77,5 +86,33 @@ export const accountList = createSelector(
         });
       })
       .toJS();
+  }
+);
+
+export const nameList = createSelector(
+  names,
+  identities,
+  (names, identities) => {
+    const identityList = [];
+    const nameList = [];
+    const nameMap = names.toObject();
+    for (let name in nameMap) {
+      const identityId = names[name];
+      let identity;
+      if (identityList.indexOf(identityId) > -1) {
+        identity = identityList.indexOf(identity);
+      } else {
+        identityList.push(identityId);
+        identity = identityList.length;
+      }
+
+      nameList.push({
+        identity,
+        name,
+        balance: identities.getIn([identityId, "balance"], 0)
+      });
+    }
+
+    return nameList;
   }
 );